refactor: runtime utilities & runtime feature enum
