+++ crates/pop-cli/src/commands/bench/block.rs
@@ -5,6 +5,7 @@ use crate::{
 	common::{
 		builds::{ensure_node_binary_exists, guide_user_to_select_profile},
 		prompt::display_message,
+		runtime::RuntimeFeature,
 	},
 };
 use clap::Args;
@@ -47,7 +48,7 @@ impl BenchmarkBlock {
 			cli,
 			target_path,
 			self.profile.as_ref().ok_or_else(|| anyhow::anyhow!("No profile provided"))?,
-			vec!["runtime-benchmarks"],
+			vec![RuntimeFeature::Benchmark.as_ref()],
 		)?;
 
 		cli.warning("NOTE: this may take some time...")?;
