+++ crates/pop-cli/src/commands/bench/storage.rs
@@ -6,6 +6,7 @@ use crate::{
 		bench::overwrite_weight_dir_command,
 		builds::{ensure_node_binary_exists, guide_user_to_select_profile},
 		prompt::display_message,
+		runtime::RuntimeFeature,
 	},
 };
 use clap::Args;
@@ -49,7 +50,7 @@ impl BenchmarkStorage {
 			cli,
 			target_path,
 			self.profile.as_ref().ok_or_else(|| anyhow::anyhow!("No profile provided"))?,
-			vec!["runtime-benchmarks"],
+			vec![RuntimeFeature::Benchmark.as_ref()],
 		)?;
 
 		cli.warning("NOTE: this may take some time...")?;
